% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dane_uzup.R
\docType{data}
\name{dane_uzup}
\alias{dane_uzup}
\title{Siatka}
\format{
Obiekt stars zawierajacy 96 wierszy i 127 kolumn oraz cztery zmienne:
\itemize{
  \item srtm srtm
  \item clc clc
  \item ndvi ndvi
  \item savi savi
}
}
\usage{
data(dane_uzup)
}
\description{
Siatka badanego obszaru.
}
\examples{
library(stars)
library(geostatbook)
data(dane_uzup)
dim(dane_uzup)
dane_uzup

}
\references{
\itemize{
  \item \url{http://srtm.csi.cgiar.org}
  \item \url{http://www.eea.europa.eu/data-and-maps}
  \item \url{http://landsat-pds.s3.amazonaws.com/}
}
}
\keyword{datasets}
