% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geostatbook-package.R
\docType{package}
\name{geostatbook-package}
\alias{geostatbook}
\alias{geostatbook-package}
\title{geostatbook: Geostatystyka w R}
\description{
Materialy do skryptu Geostatystyka w R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bookdown.org/nowosad/Geostatystyka/}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com}

}
\keyword{internal}
