% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granica.R
\docType{data}
\name{granica}
\alias{granica}
\title{Granica}
\format{SpatialPolygonsDataFrame zawierajacy obszar Suwalskiego Parku Krajobrazowego}
\usage{
granica
}
\description{
Granica badanego obszaru.
}
\examples{
library('geostatbook')
data(granica)
dim(granica)
summary(granica)
plot(granica)
}
\references{
\url{http://www.gdos.gov.pl/dane-i-metadane}
}
\keyword{datasets}

