% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punkty_ndvi.R
\docType{data}
\name{punkty_ndvi}
\alias{punkty_ndvi}
\title{Dane punktowe NDVI}
\format{SpatialPointsDataFrame zawierajacy 1011 obserwacje i 1 zmienna
\itemize{
  \item ndvi  ndvi.
}}
\usage{
punkty_ndvi
}
\description{
Zbior danych punktowych NDVI.
}
\examples{
dim(punkty_ndvi)
head(punkty_ndvi)
summary(punkty_ndvi)
}
\keyword{datasets}

