% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punkty_pref.R
\docType{data}
\name{punkty_pref}
\alias{punkty_pref}
\title{Dane punktowe preferencyjne}
\format{SpatialPointsDataFrame zawierajacy 264 obserwacje i 5 zmiennych
\itemize{
  \item srtm  srtm.
  \item clc   clc.
  \item temp  temperatura.
  \item ndvi  ndvi.
  \item savi  savi.
}}
\usage{
punkty_pref
}
\description{
Zbior danych punktowych o probkowaniu preferencyjnym.
}
\examples{
library(sf)
library(geostatbook)
data(punkty_pref)
dim(punkty_pref)
head(punkty_pref)
summary(punkty_pref)
}
\references{
\itemize{
  \item \url{http://srtm.csi.cgiar.org}
  \item \url{http://www.eea.europa.eu/data-and-maps}
  \item \url{http://landsat-pds.s3.amazonaws.com/}
}
}
\keyword{datasets}
