% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_trans.R
\name{rev_trans}
\alias{rev_trans}
\title{Transformacja odwrotna}
\usage{
rev_trans(pred, var, obs)
}
\arguments{
\item{pred}{estymowane wartosci, ktorych jednostka
jest logarytm oryginalnej jednostki}

\item{var}{wariancja kolejnych estymacji}

\item{obs}{wartosci pomiarow w punktach 
w oryginalnej jednostce}
}
\value{
wartosci po transformacji odwrotnej
}
\description{
Transformacja odwrotna (Yamamoto, 2007)
}
\references{
Yamamoto, J.K. On unbiased backtransform 
of lognormal kriging estimates. 
Comput Geosci 11, 219-234 (2007).
}
