% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punkty.R
\docType{data}
\name{punkty}
\alias{punkty}
\title{Dane punktowe}
\format{
Obiekt sf zawierajacy 242 obserwacje i 5 zmiennych
\itemize{
  \item srtm  srtm.
  \item clc   clc.
  \item temp  temperatura.
  \item ndvi  ndvi.
  \item savi  savi.
}
}
\usage{
data(punkty)
}
\description{
Zbior danych punktowych.
}
\examples{
library(sf)
library(geostatbook)
data(punkty)
head(punkty)
summary(punkty)

}
\references{
\itemize{
  \item \url{http://srtm.csi.cgiar.org}
  \item \url{http://www.eea.europa.eu/data-and-maps}
  \item \url{http://landsat-pds.s3.amazonaws.com/}
}
}
\keyword{datasets}
