% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punkty.R
\docType{data}
\name{punkty}
\alias{punkty}
\title{Dane punktowe}
\format{SpatialPointsDataFrame zawierajacy 252 obserwacje i 5 zmiennych
\itemize{
  \item srtm  srtm.
  \item clc   clc.
  \item temp  temperatura.
  \item ndvi  ndvi.
  \item savi  savi.
}}
\usage{
punkty
}
\description{
Zbior danych punktowych.
}
\examples{
library('geostatbook')
data(punkty)
dim(punkty)
head(punkty)
summary(punkty)
}
\keyword{datasets}

