% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punkty_ndvi.R
\docType{data}
\name{punkty_ndvi}
\alias{punkty_ndvi}
\title{Pollen count of Alnus, Betula, and Corylus in given polish sites}
\format{data.frame}
\description{
aerobiologia contain data about pollen count of  Alnus, Betula, 
and Corylus in given polish sites (without incomplete data from Gdańsk in years 1998 and 1999)
}

