% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punkty_ndvi.R
\docType{data}
\name{punkty_ndvi}
\alias{punkty_ndvi}
\title{Dane punktowe NDVI}
\format{SpatialPointsDataFrame zawierajacy 1011 obserwacji i 1 zmienna
\itemize{
  \item ndvi  ndvi.
}}
\usage{
punkty_ndvi
}
\description{
Zbior danych punktowych NDVI.
}
\examples{
library('geostatbook')
data(punkty_ndvi)
dim(punkty_ndvi)
head(punkty_ndvi)
summary(punkty_ndvi)
}
\references{
\itemize{
  \item \url{http://landsat-pds.s3.amazonaws.com/}
}
}
\keyword{datasets}

