% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siatka.R
\docType{data}
\name{siatka}
\alias{siatka}
\title{Siatka}
\format{SpatialPixelsDataFrame zawierajacy 127 wierszy i 96 kolumn oraz 4 zmienne
\itemize{
  \item srtm  srtm.
  \item clc   clc.
  \item ndvi  ndvi.
  \item savi  savi.
}}
\usage{
siatka
}
\description{
Siatka badanego obszaru.
}
\examples{
library('geostatbook')
data(siatka)
dim(siatka)
head(siatka)
summary(siatka)
}
\keyword{datasets}

